requirements.txt
numpy==1.22.3
Pillow==9.1.0
tflite-runtime @ file:///home/NEST1/g3-osi-sw/wheels/tflite_runtime-2.9.0-cp39-cp39-linux_armv6l.whl
picamera==1.13